Add files via upload
These files are taken from other website sources, and practicing my Python network coding skills. 
The webbrowserrefresh.py I created specifically in order refresh a browser 1000 times, this was a for a Capture the Flag, apparently, in order to successfully complete this "Hacking Capture the Flag" event, school related, this challenge required refreshing the browser 1000 times.